Copy-on-write locals and fields to protect against side effects
Previously, certain expressions (notably, binary operations) copied the
value on their left hand side to a temporary, in order to guard against
potential side effects in the RHS (e.g., a = b + (b++) would copy 'b' to
a temporary).

This both had excess overhead in a lot of cases, and also was incomplete
(function calls and array indexes did not properly copy their values to
a temporary when the RHS modified the field in question).

This change introduces explicit tracking of 'COWValues' - basically,
representing the value of a symbol at a certain point in time. If the
symbol is marked as modified (via SymbolDefinition.MarkDirty()), the
current value of the symbol is copied to a temporary, and the COWValue
updated to point to the temporary; however, in most cases, this will be
relatively rare compared to cases where this is not required.