Fix SyntaxTree name clash
The unity Visual Studio integration adds an assembly with a `SyntaxTree`
namespace; this results in a name clash with the `Microsoft.CodeAnalysis.SyntaxTree`
type which was previously implicitly referenced. This change adds using
alias directives to fix this compile error.

Note that these using aliases only work in a namespace, not at the
top-level scope, and so TestScratchpad needs to move into a namespace as
well.