Prevent Unity from complaining on scene upgrade
Unity would open a window asking you to edit a prefab if you move any component on a prefab instance if it's a added component override, which should totally be a valid operation. You can actually call the MoveComponentUp/Down on them properly, so MoveComponentRelativeToComponent is bugged or implemented inconsistently. For now we will just ignore the issue since having components ordered consistently on a behaviour is not critical since event order is technically undefined in Unity. We may want to add a pass to the pre build events that ensures consistent component ordering eventually for good measure.