Add handling for prefab instances with delta on unityengine.Object array
If a prefab instance in a scene has not been upgraded and has no delta on its string data, but a delta on its UnityEngine.Object array, it could get caught in an intermediate state where it's not upgraded fully. This adds handling to prevent that from being an issue.