Apply changes of enabled checkbox to backing behaviour
The enabled checkbox was only getting applied to backing UdonBehaviours from proxies when entering play mode and opening the inspector for a UdonSharpBehaviour. So If you changed the enabled state while in play mode to debug something, the new state would not immediately apply. Now we wrap the header drawing method and check for changes to the enabled state, then apply them to the backing UdonBehaviour. Reported by Fairplex