Fix syntax rewriter again
Trivia is pain :'(
Now we insert the attribute and either append the trailing trivia from the previously last attribute if there was an attribute, or append the leading trivia for the field declaration if it did not have attributes before.
Also add exception for VRCPlayerApi because it's a common type that gets caught by this check, but does not get serialized. Proper handling would involve checking if a type is serializable, which is not super straightforward.